requirements.txt
crewai==0.28.0
crewai-tools==0.1.13
selenium==4.15.2
psycopg2-binary==2.9.9
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.4.2 